requirements.txt
fastapi==0.115.12
pydantic-settings==2.9.1
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
jinja2==3.1.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
minio==7.2.15
numpy==2.0.2